Moved around some files and updated build steps so e2e tests pass
