+++ src/tsconfig.app.json
@@ -6,6 +6,7 @@
     "types": []
   },
   "exclude": [
+    "./util/test/*",
     "src/test.ts",
     "**/*.spec.ts"
   ]
